+++ include/math/Vector4.hpp
@@ -5,27 +5,27 @@ class Vector4 {
     public:
         union {
             struct {
-                double x, y, z, w;
+                float x, y, z, w;
             };
-            double data[4];
+            float data[4];
         };
 
         Vector4() = default;
        ~Vector4() = default;
 
        
-        Vector4(double x, double y, double z);
-        Vector4(double x, double y, double z, double w);
+        Vector4(float x, float y, float z);
+        Vector4(float x, float y, float z, float w);
         
         Vector4    operator+(const Vector4& rhs) const;
         Vector4    operator-(const Vector4& rhs) const;
         Vector4    operator-() const;
-        Vector4    operator*(double t) const;
-        double     operator*(const Vector4& rhs) const; //dot product
-        Vector4    operator/(double t) const;
+        Vector4    operator*(float t) const;
+        float     operator*(const Vector4& rhs) const; //dot product
+        Vector4    operator/(float t) const;
         Vector4    operator|(const Vector4&   v) const;
         bool      operator==(const Vector4&   v) const;
-        double        length() const;
+        float        length() const;
         void       normalize();
        
         friend std::ostream& operator<<(std::ostream& stream, const Vector4& mtx); //print Matrix4 
@@ -34,6 +34,6 @@ class Vector4 {
 };
 
 
-Vector4 operator*(double, const Vector4& rhs);
-Vector4 operator/(double, const Vector4& rhs);
+Vector4 operator*(float, const Vector4& rhs);
+Vector4 operator/(float, const Vector4& rhs);
