+++ src/camera/PinHole.cpp
@@ -2,7 +2,7 @@
 #include <cmath>
 
 
-PinHole::PinHole(double fov, double ratio, double near, double far):
+PinHole::PinHole(float fov, float ratio, float near, float far):
     fov(fov),
     asp_ratio(ratio),
     near(near),
@@ -10,7 +10,7 @@ PinHole::PinHole(double fov, double ratio, double near, double far):
 {}
 
 Matrix4 PinHole::projection_matrix() const{
-    double m[16] = {
+    float m[16] = {
                     1 / (asp_ratio * tan(fov / 2)), 0, 0, 0,
                     0, 1 / tan(fov / 2), 0, 0,
                     0, 0, (far + near) / (far - near), (2 * far * near) / (far - near),
