+++ src/material/color.cpp
@@ -1,10 +1,10 @@
 #include "../../include/material/Color.hpp"
 #include <cmath>
 
-Color::Color(const double c)
+Color::Color(const float c)
     : r(c),g(c),b(c) {}
 
-Color::Color(const double red, const double green, const double blue)
+Color::Color(const float red, const float green, const float blue)
     : r(red),g(green),b(blue) {} 
 
 
@@ -30,7 +30,7 @@ Color Color::operator*(const Color& rhs) const{
     return Color(this->r * rhs.r, this->g * rhs.g, this->b * rhs.b);
 }
 
-Color Color::operator*(const double rhs) const {
+Color Color::operator*(const float rhs) const {
     /* this->r *= rhs; */
     /* this->g *= rhs; */
     /* this->b *= rhs; */
@@ -44,7 +44,7 @@ Color operator*(const float a, const Color& c){
     return (c * a);
 }
 
-Color Color::operator/(const double rhs) const {
+Color Color::operator/(const float rhs) const {
     /* this->r *= rhs; */
     /* this->g *= rhs; */
     /* this->b *= rhs; */
@@ -58,7 +58,7 @@ bool Color::operator==(const Color& rhs) {
 }
 
 
-Color Color::powc(double p) const {
+Color Color::powc(float p) const {
     return Color(powf(r, p), powf(g, p), powf(b, p));
 }
