+++ src/objects/shapes/Sphere.cpp
@@ -1,7 +1,7 @@
 #include "../../../include/objects/shapes/Sphere.hpp"
 #include <cmath>
 
-Sphere::Sphere(const Vector4 &center, const double radius)
+Sphere::Sphere(const Vector4 &center, const float radius)
     : center(center), radius(radius) {}
 
 
@@ -10,11 +10,11 @@ Mesh Sphere::generate_mesh() const {
 
     for (unsigned int i = 0; i < (u-1); i++) {
         for (unsigned int j = 0; j < (v-1); j++) {
-            double theta = i * 180.0 / (u - 1.0);
-            double phi =   j * 360.0 / (v - 1.0);
+            float theta = i * 180.0 / (u - 1.0);
+            float phi =   j * 360.0 / (v - 1.0);
 
-            double theta1 = (i + 1.0) * 180.0 / (u - 1.0);
-            double phi1 =   (j + 1.0) * 360.0 / (v - 1.0);
+            float theta1 = (i + 1.0) * 180.0 / (u - 1.0);
+            float phi1 =   (j + 1.0) * 360.0 / (v - 1.0);
 
             Vector4 p0 = position(radius, theta, phi);
             Vector4 n0(p0.x/radius, p0.y/radius, p0.z/radius);
@@ -44,10 +44,10 @@ Mesh Sphere::generate_mesh() const {
     return mesh;
 }
 
-Vector4 Sphere::position(double radius, double theta, double phi) {
-    double x =  radius * std::sin(theta) * std::cos(phi);
-    double y =  radius * std::cos(theta);
-    double z = -radius * std::sin(theta) * std::sin(phi);
+Vector4 Sphere::position(float radius, float theta, float phi) {
+    float x =  radius * std::sin(theta) * std::cos(phi);
+    float y =  radius * std::cos(theta);
+    float z = -radius * std::sin(theta) * std::sin(phi);
 
     return Vector4(x, y, z);
 }
